requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
cv==1.0.0
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.23.4
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
keyboard==0.13.5
MarkupSafe==2.1.5
mkl==2021.4.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.0.3
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyKakao==0.0.7
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sympy==1.12.1
tqdm==4.66.4
urllib3==2.2.2